Prevent Compiler Paradox
