+++ Assets/Mirror/CompilerSymbols/Mirror.CompilerSymbols.asmdef
@@ -1,0 +1,3 @@
+﻿{
+	"name": "Mirror.CompilerSymbols"
+}
