+++ Assets/Mirror/CompilerSymbols/Mirror.CompilerSymbols.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 325984b52e4128546bc7558552f8b1d2
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
