+++ Assets/Mirror/CompilerSymbols/PreprocessorDefine.cs
@@ -1,0 +1,25 @@
+using System.Collections.Generic;
+using UnityEditor;
+
+namespace Mirror
+{
+    static class PreprocessorDefine
+    {
+        /// <summary>
+        /// Add define symbols as soon as Unity gets done compiling.
+        /// </summary>
+        [InitializeOnLoadMethod]
+        static void AddDefineSymbols()
+        {
+            HashSet<string> defines = new HashSet<string>(PlayerSettings.GetScriptingDefineSymbolsForGroup(EditorUserBuildSettings.selectedBuildTargetGroup).Split(';'))
+            {
+                "MIRROR",
+                "MIRROR_1726_OR_NEWER",
+                "MIRROR_3_0_OR_NEWER",
+                "MIRROR_3_12_OR_NEWER",
+                "MIRROR_4_0_OR_NEWER"
+            };
+            PlayerSettings.SetScriptingDefineSymbolsForGroup(EditorUserBuildSettings.selectedBuildTargetGroup, string.Join(";", defines));
+        }
+    }
+}
