+++ Assets/Mirror/CompilerSymbols/PreprocessorDefine.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 42760e6965bea614ca5706a6ba059531
+guid: f1d66fe74ec6f42dd974cba37d25d453
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
