+++ Assets/Mirror/Editor/PreprocessorDefine.cs
@@ -1,25 +1,4 @@
-using System.Collections.Generic;
-using UnityEditor;
-
-namespace Mirror
-{
-    static class PreprocessorDefine
-    {
-        /// <summary>
-        /// Add define symbols as soon as Unity gets done compiling.
-        /// </summary>
-        [InitializeOnLoadMethod]
-        static void AddDefineSymbols()
-        {
-            HashSet<string> defines = new HashSet<string>(PlayerSettings.GetScriptingDefineSymbolsForGroup(EditorUserBuildSettings.selectedBuildTargetGroup).Split(';'))
-            {
-                "MIRROR",
-                "MIRROR_1726_OR_NEWER",
-                "MIRROR_3_0_OR_NEWER",
-                "MIRROR_3_12_OR_NEWER",
-                "MIRROR_4_0_OR_NEWER"
-            };
-            PlayerSettings.SetScriptingDefineSymbolsForGroup(EditorUserBuildSettings.selectedBuildTargetGroup, string.Join(";", defines));
-        }
-    }
-}
+﻿// This file was moved to Mirror/CompilerSymbols in Mirror 4.0
+// The purpose of this file is to get the old file overwritten
+// when users update from the asset store to prevent a flood of errors
+// from having the old file still in the project as a straggler.
