+++ Assets/Mirror/Runtime/Mirror.asmdef
@@ -1,8 +1,14 @@
 {
     "name": "Mirror",
-    "references": [],
+    "references": [
+        "Mirror.CompilerSymbols"
+    ],
     "optionalUnityReferences": [],
     "includePlatforms": [],
     "excludePlatforms": [],
-    "allowUnsafeCode": false
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": []
 }
